+++ client/src/main.tsx
@@ -21,16 +21,6 @@ async function init() {
   inject();
   const steps: StepType[] = [
     {
-      selector: ".realm-selector",
-      content: "Navigate to your Realms here to manage your territories.",
-      position: "bottom",
-    },
-    {
-      selector: ".world-selector",
-      content: "Explore the World map and discover new areas.",
-      position: "bottom",
-    },
-    {
       selector: ".military-selector",
       content: "Access your Military options to manage troops and defenses.",
       position: "bottom",
@@ -78,7 +68,7 @@ async function init() {
   ];
   root.render(
     <React.StrictMode>
-      <TourProvider
+      {/* <TourProvider
         steps={steps}
         styles={{
           popover: (base) => ({
@@ -94,11 +84,11 @@ async function init() {
           controls: (base) => ({ ...base, marginTop: 100, color: "#F3C99F" }),
           close: (base) => ({ ...base, right: "auto", left: 8, top: 8 }),
         }}
-      >
-        <DojoProvider value={setupResult}>
-          <App />
-        </DojoProvider>
-      </TourProvider>
+      > */}
+      <DojoProvider value={setupResult}>
+        <App />
+      </DojoProvider>
+      {/* </TourProvider> */}
     </React.StrictMode>,
   );
 }
